refactor(pages): components/page配下のファイルをindex.tsでexportする
