+++ src/components/page/members/index.ts
@@ -1,0 +1,1 @@
+export * from './Members';
