+++ src/components/page/user-registration/index.ts
@@ -1,0 +1,1 @@
+export * from './UserRegistration';
