+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { Top } from 'src/components/page/top/Top';
+import { Top } from 'src/components/page/top';
 
 const TopPage: NextPage = () => {
   return <Top />;
