+++ src/pages/login/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { Login } from 'src/components/page/login/Login';
+import { Login } from 'src/components/page/login';
 
 const LoginPage: NextPage = () => {
   return <Login />;
