+++ src/pages/members/index.tsx
@@ -1,5 +1,5 @@
 import { NextPage } from 'next';
-import { Members } from 'src/components/page/members/Members';
+import { Members } from 'src/components/page/members';
 
 const MembersPage: NextPage = () => {
   return <Members />;
