+++ src/pages/user-registration/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { UserRegistration } from 'src/components/page/user-registration/UserRegistration';
+import { UserRegistration } from 'src/components/page/user-registration';
 
 const UserRegistrationPage: NextPage = () => {
   return <UserRegistration />;
